+++ External Cocos Helper Android Frameworks/src/sonar/systems/framework/SonarFrameworkSettings.java
@@ -6,6 +6,8 @@ public class SonarFrameworkSettings
 	
 	public static final boolean USE_GOOGLE_PLAY_GAME_SERVICES = false;	//Requires google play services library 
 	
+	public static final boolean USE_GOOGLE_ANALYTICS = false;			//Requires google play services library 
+	
 	public static final boolean USE_ADMOB = false;						//Requires  Google play services  library
 	
 	public static final boolean USE_REVMOB = false;						//Requires revmob jar file and Google play services  library
@@ -16,4 +18,5 @@ public class SonarFrameworkSettings
 	
 	public static final boolean USE_CHARTBOOST = false; 				//Requires  Google play services  library and chartboost jar
 	
+	public static final boolean USE_MOPUB = true;						//Requires MoPub SDK
 }